Moved components to own files
