+++ src/components/Reset.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Reset({onClick}: { onClick: React.EventHandler<React.MouseEvent<HTMLInputElement>> }) {
+  return <input type="reset" value="Reset" onClick={onClick}/>
+}
